+++ cmd/terraform.go
@@ -7,17 +7,17 @@ import (
 	"strings"
 )
 
-func parseInputVars(format string, input string) (string, string, string, string, error) {
+func parseInputVars(format string, input string) (string, string, string, string, string, error) {
 	if format == varFormatHCL {
 		return parseInputVarsHCL(input)
 	}
 	if format == varFormatJSON {
 		return parseInputVarsJSON(input)
 	}
-	return "", "", "", "", errors.New(`unknown var format: "` + format + `"`)
+	return "", "", "", "", "", errors.New(`unknown var format: "` + format + `"`)
 }
 
-func parseInputVarsJSON(input string) (string, string, string, string, error) {
+func parseInputVarsJSON(input string) (string, string, string, string, string, error) {
 	var data map[string]interface{}
 
 	err := json.Unmarshal([]byte(input), &data)
@@ -27,29 +27,31 @@ func parseInputVarsJSON(input string) (string, string, string, string, error) {
 	environment := data["environment"].(string)
 	profile := data["aws_profile"].(string)
 	region := data["region"].(string)
+	containerPort := data["container_port"].(string)
 
 	//did we find it?
 	if app == "" {
-		return "", "", "", "", errors.New(`missing variable: "app"`)
+		return "", "", "", "", "", errors.New(`missing variable: "app"`)
 	}
 	if environment == "" {
-		return "", "", "", "", errors.New(`missing variable: "environment"`)
+		return "", "", "", "", "", errors.New(`missing variable: "environment"`)
 	}
 	if profile == "" {
-		return "", "", "", "", errors.New(`missing variable: "profile"`)
+		return "", "", "", "", "", errors.New(`missing variable: "profile"`)
 	}
 	if region == "" {
-		return "", "", "", "", errors.New(`missing variable: "region"`)
+		return "", "", "", "", "", errors.New(`missing variable: "region"`)
 	}
 
-	return app, environment, profile, region, nil
+	return app, environment, profile, region, containerPort, nil
 }
 
-func parseInputVarsHCL(tf string) (string, string, string, string, error) {
+func parseInputVarsHCL(tf string) (string, string, string, string, string, error) {
 	app := ""
 	environment := ""
 	profile := ""
 	region := ""
+	containerPort := ""
 
 	//look for variables
 	lines := strings.Split(tf, "\n")
@@ -57,7 +59,7 @@ func parseInputVarsHCL(tf string) (string, string, string, string, error) {
 	for _, line := range lines {
 		trimmed := strings.TrimSpace(line)
 		//ignore whitespace and comments
-		if trimmed == "" || strings.HasPrefix(trimmed, "//") {
+		if trimmed == "" || strings.HasPrefix(trimmed, "//") || strings.HasPrefix(trimmed, "#") {
 			continue
 		}
 		if trimmed == "tags = {" {
@@ -75,7 +77,14 @@ func parseInputVarsHCL(tf string) (string, string, string, string, error) {
 		if len(parts) == 2 {
 			key := strings.TrimSpace(parts[0])
 			value := strings.TrimSpace(parts[1])
+
+			//remove quotes (app = "foo")
 			value = strings.Replace(value, `"`, "", -1)
+
+			//remove trailing comments (app = "foo" #comment)
+			comments := strings.Split(value, "#")
+			value = strings.TrimSpace(comments[0])
+
 			if key == "app" {
 				app = value
 			}
@@ -88,24 +97,27 @@ func parseInputVarsHCL(tf string) (string, string, string, string, error) {
 			if key == "region" {
 				region = value
 			}
+			if key == "container_port" {
+				containerPort = value
+			}
 		}
 	}
 
 	//did we find it?
 	if app == "" {
-		return "", "", "", "", errors.New(`missing variable: "app"`)
+		return "", "", "", "", "", errors.New(`missing variable: "app"`)
 	}
 	if environment == "" {
-		return "", "", "", "", errors.New(`missing variable: "environment"`)
+		return "", "", "", "", "", errors.New(`missing variable: "environment"`)
 	}
 	if profile == "" {
-		return "", "", "", "", errors.New(`missing variable: "profile"`)
+		return "", "", "", "", "", errors.New(`missing variable: "profile"`)
 	}
 	if region == "" {
-		return "", "", "", "", errors.New(`missing variable: "region"`)
+		return "", "", "", "", "", errors.New(`missing variable: "region"`)
 	}
 
-	return app, environment, profile, region, nil
+	return app, environment, profile, region, containerPort, nil
 }
 
 func updateTerraformBackend(tf string, profile string, app string, env string) string {