+++ cmd/terraform_test.go
@@ -52,6 +52,8 @@ func TestParseInputVars(t *testing.T) {
 	app = "my-app"
 	
 	environment = "qa"
+
+	container_port = "8080"
 	
 	tags = {
 		application   = ""
@@ -63,7 +65,119 @@ func TestParseInputVars(t *testing.T) {
 	
 	internal = "true"
 `
-	app, env, profile, region, err := parseInputVars(varFormatHCL, tf)
+	app, env, profile, region, containerPort, err := parseInputVars(varFormatHCL, tf)
+
+	t.Log(app)
+	t.Log(env)
+	t.Log(profile)
+	t.Log(region)
+	t.Log(containerPort)
+	if err != nil {
+		t.Error(err)
+	}
+
+	expected := "my-app"
+	if app != expected {
+		t.Errorf("expected: %s; actual: %s", expected, app)
+	}
+	expected = "qa"
+	if env != expected {
+		t.Errorf("expected: %s; actual: %s", expected, env)
+	}
+	expected = "default"
+	if profile != expected {
+		t.Errorf("expected: %s; actual: %s", expected, profile)
+	}
+	expected = "us-east-2"
+	if region != expected {
+		t.Errorf("expected: %s; actual: %s", expected, profile)
+	}
+	expected = "8080"
+	if containerPort != expected {
+		t.Errorf("expected: %s; actual: %s", containerPort, profile)
+	}
+}
+
+func TestParseInputVars_OptionalContainerPort(t *testing.T) {
+
+	tf := `
+	region = "us-east-2"
+
+	aws_profile = "default"
+	
+	saml_role = "devops"
+	
+	app = "my-app"
+	
+	environment = "qa"
+
+	tags = {
+		application   = ""
+		environment   = "dev"
+		team          = ""
+		customer      = ""
+		contact-email = ""
+	}
+	
+	internal = "true"
+`
+	app, env, profile, region, containerPort, err := parseInputVars(varFormatHCL, tf)
+
+	t.Log(app)
+	t.Log(env)
+	t.Log(profile)
+	t.Log(region)
+	t.Log(containerPort)
+	if err != nil {
+		t.Error(err)
+	}
+
+	expected := "my-app"
+	if app != expected {
+		t.Errorf("expected: %s; actual: %s", expected, app)
+	}
+	expected = "qa"
+	if env != expected {
+		t.Errorf("expected: %s; actual: %s", expected, env)
+	}
+	expected = "default"
+	if profile != expected {
+		t.Errorf("expected: %s; actual: %s", expected, profile)
+	}
+	expected = "us-east-2"
+	if region != expected {
+		t.Errorf("expected: %s; actual: %s", expected, profile)
+	}
+	expected = ""
+	if containerPort != expected {
+		t.Errorf("expected: %s; actual: %s", containerPort, profile)
+	}
+}
+
+func TestParseInputVars_comments(t *testing.T) {
+
+	tf := `
+	region = "us-east-2"
+
+	aws_profile = "default"
+	
+	saml_role = "devops"
+	
+	app = "my-app" #comment
+	
+	environment = "qa"
+	
+	tags = {
+		application   = ""
+		environment   = "dev"
+		team          = ""
+		customer      = ""
+		contact-email = ""
+	}
+	
+	internal = "true"
+`
+	app, env, profile, region, _, err := parseInputVars(varFormatHCL, tf)
 
 	t.Log(app)
 	t.Log(env)
@@ -99,6 +213,7 @@ func TestParseInputVars_JSON(t *testing.T) {
 	"saml_role": "devops",	
 	"app": "my-app",	
 	"environment": "qa",	
+	"container_port": "8080",
 	"tags": {
 		"application": "",
 		"environment": "dev",
@@ -109,12 +224,13 @@ func TestParseInputVars_JSON(t *testing.T) {
 	"internal": true
 }
 `
-	app, env, profile, region, err := parseInputVars(varFormatJSON, tf)
+	app, env, profile, region, containerPort, err := parseInputVars(varFormatJSON, tf)
 
 	t.Log(app)
 	t.Log(env)
 	t.Log(profile)
 	t.Log(region)
+	t.Log(containerPort)
 	if err != nil {
 		t.Error(err)
 	}
@@ -135,4 +251,9 @@ func TestParseInputVars_JSON(t *testing.T) {
 	if region != expected {
 		t.Errorf("expected: %s; actual: %s", expected, profile)
 	}
+	expected = "8080"
+	if containerPort != expected {
+		t.Errorf("expected: %s; actual: %s", expected, profile)
+	}
+
 }