+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.3.0 (2018-09-20)
+
+### Features
+
+- Sets optional port in docker-compose.yml ([#13](https://github.com/turnerlabs/fargate-create/issues/13))
+
+### Bug Fixes
+
+  - Fixes issue parsing Terraform comments ([#12](https://github.com/turnerlabs/fargate-create/issues/12))
+
+
 ## 0.2.1 (2018-09-12)
 
 ### Features